refactor: cleanup svg imports
