+++ src/assets/icons/index.tsx
@@ -1,0 +1,33 @@
+import { ReactComponent as AppleIcon } from "./svg/Apple.svg";
+import { ReactComponent as PineappleIcon } from "./svg/Pineapple.svg";
+import { ReactComponent as StrawberryIcon } from "./svg/Strawberry.svg";
+import { ReactComponent as BananaIcon } from "./svg/Banana.svg";
+import { ReactComponent as AvocadoIcon } from "./svg/Avocado.svg";
+import { ReactComponent as MelonIcon } from "./svg/Melon.svg";
+import { ReactComponent as GrapesIcon } from "./svg/Grape.svg";
+import { ReactComponent as DayIcon } from "./svg/Day.svg";
+import { ReactComponent as AppleAvatar } from './svg/AppleAvatar.svg';
+import { ReactComponent as AvocadoAvatar } from './svg/AvocadoAvatar.svg';
+import { ReactComponent as CoconutAvatar } from './svg/CoconutAvatar.svg';
+import { ReactComponent as LemonAvatar } from './svg/LemonAvatar.svg';
+import { ReactComponent as OrangeAvatar } from './svg/OrangeAvatar.svg';
+import { ReactComponent as PearAvatar } from './svg/PearAvatar.svg';
+import { ReactComponent as AppLogo } from './svg/smartsystemlogo.svg';
+
+export { 
+      AppleIcon, 
+      PineappleIcon,  
+      StrawberryIcon,
+      BananaIcon,
+      AvocadoIcon,
+      MelonIcon,
+      GrapesIcon,
+      DayIcon,
+      AppleAvatar,
+      AvocadoAvatar,
+      CoconutAvatar,
+      LemonAvatar,
+      OrangeAvatar,
+      PearAvatar,
+      AppLogo,
+    }
